requirements.txt
gunicorn==20.0.4
PyNaCl==1.5.0
requests~=2.28.1
fire~=0.5.0
Flask~=2.2.2
jsonpickle==3.0.1